+++ src/Controller/PatchElementController.php
@@ -3,11 +3,13 @@
 namespace App\Controller;
 
 use App\Exception\ClientNotFoundException;
+use App\Exception\ClientUnauthorizedException;
 use App\Helper\Regex;
 use App\Response\NoContentResponse;
+use App\Security\AccessChecker;
 use App\Security\AuthProvider;
-use App\Security\PermissionChecker;
 use App\Service\ElementManager;
+use App\Type\AccessType;
 use Ramsey\Uuid\Rfc4122\UuidV4;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -19,7 +21,7 @@ class PatchElementController extends AbstractController
     public function __construct(
         private ElementManager $elementManager,
         private AuthProvider $authProvider,
-        private PermissionChecker $permissionChecker
+        private AccessChecker $accessChecker
     ) {
     }
 
@@ -34,24 +36,22 @@ class PatchElementController extends AbstractController
     public function patchElement(string $uuid, Request $request): Response
     {
         $elementUuid = UuidV4::fromString($uuid);
-        $hasPermission = $this->permissionChecker->checkPermissionToElement(
-            $this->authProvider->getUserUuid(),
-            $elementUuid,
-            'WRITE'
-        );
-        if (!$hasPermission) {
-            throw new ClientNotFoundException();
+        $userUuid = $this->authProvider->getUserUuid();
+
+        if (!$userUuid) {
+            throw new ClientUnauthorizedException();
         }
 
-        $data = $request->getContent();
-        $data = \Safe\json_decode($data, true);
+        if (!$this->accessChecker->hasAccessToElement($userUuid, $elementUuid, AccessType::UPDATE)) {
+            throw new ClientNotFoundException();
+        }
 
         $element = $this->elementManager->getElement($elementUuid);
         if (null === $element) {
             throw new ClientNotFoundException();
         }
-        $element->addProperties($data);
-        $this->elementManager->merge($element);
+        $data = \Safe\json_decode($request->getContent(), true);
+        $this->elementManager->mergeWithUserDefinedData($element, $data);
         $this->elementManager->flush();
 
         return new NoContentResponse();