+++ src/Controller/PutElementController.php
@@ -3,11 +3,13 @@
 namespace App\Controller;
 
 use App\Exception\ClientNotFoundException;
+use App\Exception\ClientUnauthorizedException;
 use App\Helper\Regex;
 use App\Response\NoContentResponse;
+use App\Security\AccessChecker;
 use App\Security\AuthProvider;
-use App\Security\PermissionChecker;
 use App\Service\ElementManager;
+use App\Type\AccessType;
 use Ramsey\Uuid\Rfc4122\UuidV4;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -19,7 +21,7 @@ class PutElementController extends AbstractController
     public function __construct(
         private ElementManager $elementManager,
         private AuthProvider $authProvider,
-        private PermissionChecker $permissionChecker
+        private AccessChecker $accessChecker
     ) {
     }
 
@@ -34,27 +36,32 @@ class PutElementController extends AbstractController
     public function putElement(string $uuid, Request $request): Response
     {
         $elementUuid = UuidV4::fromString($uuid);
-        $hasPermission = $this->permissionChecker->checkPermissionToElement(
-            $this->authProvider->getUserUuid(),
-            $elementUuid,
-            'WRITE'
-        );
-        if (!$hasPermission) {
-            throw new ClientNotFoundException();
+        $userUuid = $this->authProvider->getUserUuid();
+
+        if (!$userUuid) {
+            throw new ClientUnauthorizedException();
         }
 
-        $data = $request->getContent();
-        $data = \Safe\json_decode($data, true);
+        if (!$this->accessChecker->hasAccessToElement($userUuid, $elementUuid, AccessType::UPDATE)) {
+            throw new ClientNotFoundException();
+        }
 
         $element = $this->elementManager->getElement($elementUuid);
         if (null === $element) {
             throw new ClientNotFoundException();
         }
         foreach ($element->getProperties() as $name => $value) {
+            if ('id' === $name) {
+                continue;
+            }
+            if ('created' === $name) {
+                continue;
+            }
             $element->addProperty($name, null);
         }
-        $element->addProperties($data);
-        $this->elementManager->merge($element);
+        $data = \Safe\json_decode($request->getContent(), true);
+
+        $this->elementManager->mergeWithUserDefinedData($element, $data);
         $this->elementManager->flush();
 
         return new NoContentResponse();